+++ lib/wax/utils/jws.ex
@@ -1,6 +1,14 @@
 defmodule Wax.Utils.JWS do
   @moduledoc false
 
+  require Record
+
+  Record.defrecord(
+    :ecdsa_signature,
+    :"ECDSA-Sig-Value",
+    Record.extract(:"ECDSA-Sig-Value", from_lib: "public_key/include/OTP-PUB-KEY.hrl")
+  )
+
   @spec verify(String.t(), binary()) :: :ok | {:error, any()}
   def verify(jws, root_cert_der) do
     try do
@@ -16,13 +24,12 @@ defmodule Wax.Utils.JWS do
         |> Enum.map(&Base.decode64!/1)
         |> Kernel.++([root_cert_der])
 
-      leaf_cert = List.first(cert_chain)
-
       case :public_key.pkix_path_validation(root_cert_der, Enum.reverse(cert_chain), []) do
         #FIXME: shall we check the CRL? Or are FIDO MDS sufficient?
         {:ok, _} ->
           public_key =
-            leaf_cert
+            cert_chain
+            |> List.first()
             |> X509.Certificate.from_der!()
             |> X509.Certificate.public_key()
 
@@ -30,10 +37,9 @@ defmodule Wax.Utils.JWS do
 
           message = header_b64 <> "." <> payload_b64
 
-          sig = Base.url_decode64!(sig_b64, padding: false)
+          File.write("testouille", message)
 
-          #FIXME: we don't check if the alg of the JWS header and the public key of
-          # the certificate match (i.e. same algorithm) -> should we?
+          sig = der_encoded_sig(sig_b64)
 
           if :public_key.verify(message, digest_alg, sig, public_key) do
             :ok
@@ -48,6 +54,28 @@ defmodule Wax.Utils.JWS do
       _ ->
         {:error, :jws_decode_error}
     end
+  end
+
+  @spec der_encoded_sig(String.t()) :: binary()
+  defp der_encoded_sig(sig_b64) do
+    Base.url_decode64!(sig_b64, padding: false)
+    |> new_der_encoded_sig()
+    |> to_der_sig()
+  end
+
+  defp new_der_encoded_sig(r, s) when is_integer(r) and is_integer(s) do
+    ecdsa_signature(r: r, s: s)
+  end
+
+  defp new_der_encoded_sig(raw) when is_binary(raw) do
+    size = raw |> byte_size() |> div(2)
+    <<r::size(size)-unit(8), s::size(size)-unit(8)>> = raw
+    new_der_encoded_sig(r, s)
+  end
+
+  # Export to DER binary format, for use with :public_key.verify/4
+  defp to_der_sig(ecdsa_signature() = signature) do
+    :public_key.der_encode(:"ECDSA-Sig-Value", signature)
   end
 
   @spec digest_alg(String.t()) :: :public_key.digest_type()